+++ src/Component/OpenApi3/Tests/fixtures/whitelisted-paths/expected/Endpoint/FindTweetsById.php
@@ -60,7 +60,7 @@ class FindTweetsById extends \Jane\Component\OpenApi3\Tests\Expected\Runtime\Cli
      * {@inheritdoc}
      *
      *
-     * @return null|\Jane\Component\OpenApi3\Tests\Expected\Model\TweetLookupResponse
+     * @return null|\Jane\Component\OpenApi3\Tests\Expected\Model\TweetLookupResponse|\Jane\Component\OpenApi3\Tests\Expected\Model\Error
      */
     protected function transformResponseBody(\Psr\Http\Message\ResponseInterface $response, \Symfony\Component\Serializer\SerializerInterface $serializer, ?string $contentType = null)
     {
@@ -70,7 +70,7 @@ class FindTweetsById extends \Jane\Component\OpenApi3\Tests\Expected\Runtime\Cli
             return $serializer->deserialize($body, 'Jane\Component\OpenApi3\Tests\Expected\Model\TweetLookupResponse', 'json');
         }
         if (mb_strpos(strtolower($contentType), 'application/json') !== false) {
-            return json_decode($body);
+            return $serializer->deserialize($body, 'Jane\Component\OpenApi3\Tests\Expected\Model\Error', 'json');
         }
         if (mb_strpos(strtolower($contentType), 'application/problem+json') !== false) {
             return json_decode($body);