+++ src/Component/OpenApi3/Tests/fixtures/whitelisted-paths/expected/Normalizer/JaneObjectNormalizer.php
@@ -18,6 +18,8 @@ class JaneObjectNormalizer implements DenormalizerInterface, NormalizerInterface
     use ValidatorTrait;
     protected $normalizers = [
         
+        \Jane\Component\OpenApi3\Tests\Expected\Model\Error::class => \Jane\Component\OpenApi3\Tests\Expected\Normalizer\ErrorNormalizer::class,
+        
         \Jane\Component\OpenApi3\Tests\Expected\Model\Expansions::class => \Jane\Component\OpenApi3\Tests\Expected\Normalizer\ExpansionsNormalizer::class,
         
         \Jane\Component\OpenApi3\Tests\Expected\Model\TweetLookupResponse::class => \Jane\Component\OpenApi3\Tests\Expected\Normalizer\TweetLookupResponseNormalizer::class,
@@ -64,6 +66,7 @@ class JaneObjectNormalizer implements DenormalizerInterface, NormalizerInterface
     {
         return [
             
+            \Jane\Component\OpenApi3\Tests\Expected\Model\Error::class => false,
             \Jane\Component\OpenApi3\Tests\Expected\Model\Expansions::class => false,
             \Jane\Component\OpenApi3\Tests\Expected\Model\TweetLookupResponse::class => false,
             \Jane\Component\OpenApi3\Tests\Expected\Model\Poll::class => false,